\name{batchBreast}
\alias{batchBreast}
\docType{data}
\title{
A vector identifying batches in the breast dataset.
}
\description{
A vector identifying batches in the breast dataset.
}
\usage{data("batchBreast")}
\format{
  The format is:
 num [1:70] 1 1 1 1 1 1 1 1 1 1 ...
}
\details{
A vector identifying batches in the breast dataset.
}
\source{
Nyamundanda, G. and Sadanandam, A., 2016. A novel and robust statistical method to diagnose and correct batch effects in large-scale genomic data.
}
\references{
Nyamundanda, G. and Sadanandam, A., 2016. A novel and robust statistical method to diagnose and correct batch effects in large-scale genomic data.
}
\examples{
data(batchBreast)
## maybe str(batchBreast) ; plot(batchBreast) ...
}
\keyword{datasets}
