\name{batchColon}
\alias{batchColon}
\docType{data}
\title{
A vector identifying batches in the colon dataset.
}
\description{
A vector identifying batches in the colon dataset.
}
\usage{data("batchColon")}
\format{
  The format is:
 num [1:112] 1 1 1 1 1 1 1 1 1 1 ...
}
\details{
A vector identifying batches in the colon dataset.
}
\source{
Nyamundanda, G. and Sadanandam, A., 2016. A novel and robust statistical method to diagnose and correct batch effects in large-scale genomic data.
}
\references{
Nyamundanda, G. and Sadanandam, A., 2016. A novel and robust statistical method to diagnose and correct batch effects in large-scale genomic data.
}
\examples{
data(batchColon)
## maybe str(batchColon) ; plot(batchColon) ...
}
\keyword{datasets}
